iiol/requirements/common.txt
django~=4.2.4
django-bootstrap5
pillow
easy-thumbnails
djangorestframework
markdown
django-filter 
python-dotenv

#barcode detect
opencv-python
pyzbar

requests


#postgreSQL driver
psycopg2-binary

#redis
django-redis

celery